+++ testutil/testnode/full_node_test.go
@@ -8,12 +8,14 @@ import (
 	"github.com/cosmos/cosmos-sdk/client"
 	"github.com/stretchr/testify/suite"
 	"github.com/tendermint/tendermint/config"
+	tmrand "github.com/tendermint/tendermint/libs/rand"
 )
 
 type IntegrationTestSuite struct {
 	suite.Suite
 
 	cleanups []func()
+	accounts []string
 	cctx     client.Context
 }
 
@@ -21,16 +23,21 @@ func (s *IntegrationTestSuite) SetupSuite() {
 	s.T().Log("setting up integration test suite")
 	require := s.Require()
 
+	// we create an arbitray number of funded accounts
+	for i := 0; i < 300; i++ {
+		s.accounts = append(s.accounts, tmrand.Str(9))
+	}
+
 	tmCfg := config.DefaultConfig()
 	tmCfg.Consensus.TimeoutCommit = time.Millisecond * 100
-	tmNode, app, cctx, err := New(s.T(), tmCfg, false, "taco", "salad")
+	tmNode, app, cctx, err := New(s.T(), tmCfg, false, s.accounts...)
 	require.NoError(err)
 
 	cctx, stopNode, err := StartNode(tmNode, cctx)
 	require.NoError(err)
 	s.cleanups = append(s.cleanups, stopNode)
 
-	cctx, cleanupGRPC, err := StartGRPCServer(app, *DefaultAppConfig(), cctx)
+	cctx, cleanupGRPC, err := StartGRPCServer(app, DefaultAppConfig(), cctx)
 	s.cleanups = append(s.cleanups, cleanupGRPC)
 
 	s.cctx = cctx
@@ -50,8 +57,9 @@ func (s *IntegrationTestSuite) Test_Liveness() {
 	// check that we're actually able to set the consensus params
 	params, err := s.cctx.Client.ConsensusParams(context.TODO(), nil)
 	require.NoError(err)
-	require.Equal(1, params.ConsensusParams.Block.TimeIotaMs)
+	require.Equal(int64(1), params.ConsensusParams.Block.TimeIotaMs)
 	_, err = WaitForHeight(s.cctx, 20)
+	require.NoError(err)
 }
 
 func TestIntegrationTestSuite(t *testing.T) {
